prueba_tecnica_python/requirements.txt
python-decouple
djangorestframework
django-cors-headers
pandas
typing-extensions